zuventure/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-ckeditor==6.7.2
django-js-asset==3.0.1
pillow==11.1.0
sqlparse==0.5.3
tzdata==2024.2
